Fix a hang when we're not on a terminal.
TIOCGWINSZ obviously works only on a tty; otherwise the return value will be
bogus.  In fact, it can be bogus even on actual terminals -- for example, a
serial console has no way to send this data and leaves the size as 0x0;
viewing an image over serial is a quite likely use of catimg (no or clumsy
ways to send the image over for graphical viewing).  You can simulate such a
case with "stty columns 0 rows 0".

A more mainstream way to trigger this hang:
    some pipe|while read x;do catimg "$x";done
which has redirected stdin.

If we're not on a tty, let's use $COLUMNS, it's set by some shells.  If
that's not set either, 80 is a reasonable default.